Exercise 2
Write a flask app where a user can enter an integer with a GET form. When the user submits the form, the page you route to should show whether the number was an even number, odd number, or not an integer at all. In this resultant page, you should have a link that brings the user back home. Because you are using a GET form, this resultant page should still work if you navigate to it directly, but with an error message if no query parameter was included. Remember to extract the common template HTML code to a layout.