Create Exercise3
Write a flask app that allows users to register for different student organizations. The home page should have a simple POST form that allows users to enter their name and then select one of five student organizations. When the user registers, they should be added to a global dictionary in the server where the key is the student's name and the value is the organization. After the user is saved, redirect to a page that shows the list of all registered users. This other page should also be accessible from the home page and vice versa via a navbar. Make sure you have both frontend and backend validation on your form to require the name and organization and only allow the 5 organizations you've hardcoded (they can be any organizations you want to hard code). Also make sure that you have extracted the common template HTML code to a layout.